pyproject.toml
[project]
name = "doc-upload-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.0",
    "fastapi>=0.129.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "openai>=2.20.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]
